move validation to match upstream convention
